Drop process.fixed_libs, enable breakpoints right away
The theory behing fixed_libs was broken from the beginning, we can't tell
dlopened libraries from fixed ones when attaching.  We also don't need to
wait for _start to enable breakpoints.  We may need this on LS_TOPLT_POINT
architectures, but that was added for PPC, and PPC doesn't need it anymore.