Add auxv reader
We should be able to instrument dynamic linker as well, if that is
requested.  Unfortunately we may still need artificial _start probe,
because we don't know _r_debug address until the dynamic linker fills
the DT_DEBUG entry.  (Though we are likely to find _r_debug in symbol
table if we look.)