Drop breakpoint->proc
This assumed that the process where this breakpoint is added can also be
used to turn the breakpoint on and off.  That's not true, leader is not
always stopped when we need it.  But when we need to turn a breakpoint on
or off, we are doing so in response to an event, and thus do have a stopped
process in hands.  Just use that.