Resolve more issues on big endian machines
We need to cover the other direction as well.  Note that the assert in
value_set_word currently fires.  That's because when we are passing structs
by value, those structs may be larger than what fits in long.  We may need
to add a heuristic to fetch.c to support legacy back ends.  When those
asserts are removed, most test suite actually passes (6 failures related
to ABI support, 2 failures in filtering test).