Introduce parameter pack flavors and parameter pack backend callbacks
The issue this is trying to solve is that some backends (Itanium in
particular) change the parameter passing convention depending on whether
given parameter is part of varargs parameter or not.  Therefore the
backend needs to know a) that any give argument is part of parameter pack,
and b) if it is, whether it's a varargs-type pack, or something else.

Backends that wish to use this should add #define ARCH_HAVE_FETCH_PACK and
define functions arch_fetch_param_pack_{start,end} (with the same interface
as fetch_param_pack_{start,end} from fetch.h).