mipsel: Improve singlestepping over branch always
When singlestepping over BEQ issuing a compare between
a single register (i.e a branch always), only follow the
branch path.

This catches the case for functions that end with loops
and avoids us putting breakpoints at the begining of the
next function (where there likely already is a bp).

Signed-off-by: Edgar E. Iglesias <edgar@axis.com>