Add proc_find_symbol, use it to find corresponding symbols for clones
The code in breakpoint_clone was buggy--it looked through old_proc's
symbols.  It was also needlessly verbose, proc_find_symbol uses library
key to find the right library.  Instead use the new interface and call
it on new_proc to find the new symbol.

We also didn't relink symbols in stack elements of a cloned process in
process_clone.  Use proc_find_symbol for this as well.