Guard number of expanded structures of the same type
The logic behind that is that if we limit array output, we should similarly
limit linked list output.  Number of expansions of the same-typed structure
is a reasonable metric for that.  Double recursive structures will expand
a bit too much, but on the other hand we want the stuff that is typed out
in the configure file to display fully, and not be considered for trimming.

Test suite included.