Convert opt_F_t to support directories as well as files, store in vector
Also allow parsing colon-separated lists into opt_F.  This will be useful
for -F as well as for parsing some environment variables.

This removes bits related to default directories to look into.  This will
eventually be moved to os_get_config_dirs, currently it's simply disabled.
The code in libltrace.c is currently incompatible with opt_F being a vector,
so it's disabled as well.  Revamp of the whole configury shebang underway.