Ditch the unbuffered IO in xinfdump
I don't believe this is honestly necessary, normal debug() calls go
through stderr stream, and stderr should be unbuffered by default
anyway.  If this is honestly important, we might consider using
dprintf instead, and go through fprintf only when that's unavailable
(it's in POSIX 2008, so relatively recent and presumably not widely
available).