Introduce a module for managing prototypes and prototype libraries
This is to be used, eventually, for keeping one prototype library per DSO.
Right now we only have one prototype library, like before: g_prototypes.
Typedef submodule is not yet used, only prototypes are stored in the
protolib as of now.

The change in Makefile.am is actually an inclusion of prototype.{c,h}
followed by a rewrap of name lists.