I now always build the export list
Previously I only built the export list when tracing with -l. Since I was using
this export list to resolve aliased symbols in addition to setting breakpoints,
this aliased symbol resolution was only working with -l. I now always build the
export list to make aliased symbol resolution always work.

I now have a separate variable to control whether we should activate latent
symbols or not; previously the existence of the export list was used to make
this determination.

Furthermore populate_this_symtab() now takes an extra argument to indicate that
ONLY the export list should be filled in