Initialize the PLT slot map correctly on x86 and x86_64
The PLT slot map translates relocation numbers to PLT slot numbers,
but was actually initialized in the opposite direction.  Fix the way
it's initialized.  This bug can be seen on glibc in particular:

  $ ltrace -e free ls
  libc.so.6->free(0x5)           = <void>
  libc.so.6->free(0x78)          = <void>
  libc.so.6->free(0xc)           = <void>
  libc.so.6->free(0x308)         = <void>

Note the nonsense values passed to free.  The problem is that these
are not free calls at all, but malloc calls that are assigned to wrong
PLT slots due to above bug.