DWARF tests now work with glibc 2.22
Different versions of glibc define their functions differently and thus generate
different debug symbols. In particular, sometime between glibc 2.19 and 2.22 a
number of function definitions were changed to be generated by a preprocessor
macro. This causes the full prototype of these functions to not be available.
This affects nanosleep() and strlen() in the tests. Thus for these tests I only
check that the function was called and not anything particular about the
arguments